MDEV-22929 MariaBackup option to report and/or continue when corruption is encountered
Post-push Windows compilation errors fix.

+++ extra/mariabackup/xtrabackup.cc
@@ -454,7 +454,7 @@ bool CorruptedPages::print_to_file(const char *filename) const
   ut_a(!pthread_mutex_unlock(&m_mutex));
   if (xtrabackup_backup)
     return backup_file_print_buf(filename, out.str().c_str(),
-                                 out.str().size());
+                                 static_cast<int>(out.str().size()));
   std::ofstream outfile;
   outfile.open(filename);
   if (!outfile.is_open())
@@ -542,7 +542,7 @@ void CorruptedPages::zero_out_free_pages()
              space_it->second.pages.begin();
          page_it != space_it->second.pages.end(); ++page_it)
     {
-      bool is_free= fseg_page_is_free(space, *page_it);
+      bool is_free= fseg_page_is_free(space, static_cast<unsigned>(*page_it));
       if (!is_free) {
         space_info_t &space_info = non_free_pages[space_id];
         space_info.pages.insert(*page_it);
